+++ src/main/java/net/geforcemods/securitycraft/compat/waila/WailaDataProvider.java
@@ -101,9 +101,8 @@ public class WailaDataProvider implements IWailaPlugin, IComponentProvider, IEnt
 		TileEntity te = data.getTileEntity();
 
 		//last part is a little cheaty to prevent owner info from being displayed on non-sc blocks
-		if (config.get(SHOW_OWNER) && te instanceof IOwnable && block.getRegistryName().getNamespace().equals(SecurityCraft.MODID)) {
+		if (config.get(SHOW_OWNER) && te instanceof IOwnable && block.getRegistryName().getNamespace().equals(SecurityCraft.MODID))
 			body.add(Utils.localize("waila.securitycraft:owner", PlayerUtils.getOwnerComponent(((IOwnable) te).getOwner().getName())));
-		}
 
 		if (disguised)
 			return;